myrequirements.txt
absl-py==1.0.0
astunparse==1.6.3
attrs==21.4.0
audioread==2.1.9
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
clldutils==3.10.1
colorlog==6.6.0
csvw==2.0.0
cycler==0.11.0
Cython==0.29.28
decorator==5.1.1
dill==0.3.4
flatbuffers==2.0
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.2
isodate==0.6.1
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
librosa==0.7.1
llvmlite==0.31.0
Markdown==3.3.6
matplotlib==3.2.2
multiprocess==0.70.12.2
numba==0.48.0
numpy==1.21.5
oauthlib==3.2.0
opt-einsum==3.3.0
p-tqdm==1.3.3
pathos==0.2.8
phonemizer==2.2.1
pox==0.3.0
ppft==1.6.6.4
protobuf==3.19.4
py-espeak-ng==0.1.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.7
python-dateutil==2.8.2
pyworld==0.3.0
regex==2022.3.2
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rfc3986==1.5.0
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.2
scipy==1.7.3
segments==2.2.0
six==1.16.0
SoundFile==0.10.3.post1
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tqdm==4.40.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
webrtcvad==2.0.10
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.7.0
